ml_backend/requirements.txt
ccxt==1.57.59
numpy==1.19.2
pandas==1.2.5
u8darts[torch]==0.14.0
pymongo==3.12.0
ta==0.7.0
dnspython==1.16.0
docopt==0.6.2
